--- src/main/java/bowling/domain/Frame.java
@@ -1,16 +1,0 @@
-package bowling.domain;
-
-public interface Frame {
-
-    void setSecondScore(Score secondScore);
-
-    String convert();
-
-    boolean isStrike();
-
-    boolean isSpare();
-
-    void setThirdScore(Score thirdScore);
-
-    boolean isLastFrame();
-}
