+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -2,32 +2,19 @@ package bowling.domain.frame;
 
 import bowling.domain.Score;
 
-public class NormalFrame extends Frame {
-
-
-    public NormalFrame() {
-    }
+public class NormalFrame implements Frame {
+    private final Score firstScore;
+    private Score secondScore;
 
     public NormalFrame(Score firstScore) {
         this.firstScore = firstScore;
     }
 
-    public NormalFrame(Score firstScore, Score secondScore) {
-        this.firstScore = firstScore;
+    public void setSecondScore(Score secondScore) {
         this.secondScore = secondScore;
     }
 
-    @Override
-    public void makeScore(Score score, int index) {
-        if (index == firstScoreKey) {
-            this.firstScore = score;
-        }
-        if (index == secondScoreKey) {
-            this.secondScore = score;
-        }
-    }
-
-    public boolean isStrike() {
+    private boolean isStrike() {
         return firstScore.equals(new Score(10));
     }
 
@@ -37,27 +24,23 @@ public class NormalFrame extends Frame {
     }
 
     @Override
+    public void setThirdScore(Score thirdScore) {
+    }
+
+    @Override
     public boolean isLastFrame() {
         return false;
     }
 
-    public boolean hasDoneFirstPitch() {
-        return firstScore != null && secondScore == null;
-    }
-
-    @Override
-    public boolean hasDoneSecondPitch() {
-        return false;
+    public boolean isFirstPitch() {
+        return secondScore == null;
     }
 
     public String convert() {
-        if (isEmpty()) {
-            return "";
-        }
         if (isStrike()) {
             return "X";
         }
-        if (hasDoneFirstPitch() && !isStrike()) {
+        if (isFirstPitch() && !isStrike()) {
             return firstScore.convert() + "|";
         }
         if (isSpare(firstScore, secondScore)) {
